Upload zipped project
Here is the current state of my project.  I implemented the list loaded in sections, but I did not have the time to implement the spinning thing while loading.  Some of the fields in the detail view take you to a list of Pokemon that share some trait.  I Implemented that functionality for those traits that had such a list in the api.